Configurable `http_timeout` on API requests  (#70)
This PR adds support for setting a timeout for executing API requests. By default, it will use 60 seconds (which is also the default RestClient uses) but this can be changed by setting `config.http_timeout = xxx` to the desired number of seconds. To support this feature, the code calling RestClient had to be fully refactored.

Fixes #63